+++ core/src/main/java/bisq/core/grpc/BisqGrpcServer.java
@@ -22,27 +22,27 @@ import bisq.core.payment.PaymentAccount;
 import bisq.core.trade.handlers.TransactionResultHandler;
 import bisq.core.trade.statistics.TradeStatistics2;
 
-import bisq.grpc.GetBalanceGrpc;
-import bisq.grpc.GetBalanceReply;
-import bisq.grpc.GetBalanceRequest;
-import bisq.grpc.GetOffersGrpc;
-import bisq.grpc.GetOffersReply;
-import bisq.grpc.GetOffersRequest;
-import bisq.grpc.GetPaymentAccountsGrpc;
-import bisq.grpc.GetPaymentAccountsReply;
-import bisq.grpc.GetPaymentAccountsRequest;
-import bisq.grpc.GetTradeStatisticsGrpc;
-import bisq.grpc.GetTradeStatisticsReply;
-import bisq.grpc.GetTradeStatisticsRequest;
-import bisq.grpc.GetVersionGrpc;
-import bisq.grpc.GetVersionReply;
-import bisq.grpc.GetVersionRequest;
-import bisq.grpc.PlaceOfferGrpc;
-import bisq.grpc.PlaceOfferReply;
-import bisq.grpc.PlaceOfferRequest;
-import bisq.grpc.StopServerGrpc;
-import bisq.grpc.StopServerReply;
-import bisq.grpc.StopServerRequest;
+import bisq.proto.grpc.GetBalanceGrpc;
+import bisq.proto.grpc.GetBalanceReply;
+import bisq.proto.grpc.GetBalanceRequest;
+import bisq.proto.grpc.GetOffersGrpc;
+import bisq.proto.grpc.GetOffersReply;
+import bisq.proto.grpc.GetOffersRequest;
+import bisq.proto.grpc.GetPaymentAccountsGrpc;
+import bisq.proto.grpc.GetPaymentAccountsReply;
+import bisq.proto.grpc.GetPaymentAccountsRequest;
+import bisq.proto.grpc.GetTradeStatisticsGrpc;
+import bisq.proto.grpc.GetTradeStatisticsReply;
+import bisq.proto.grpc.GetTradeStatisticsRequest;
+import bisq.proto.grpc.GetVersionGrpc;
+import bisq.proto.grpc.GetVersionReply;
+import bisq.proto.grpc.GetVersionRequest;
+import bisq.proto.grpc.PlaceOfferGrpc;
+import bisq.proto.grpc.PlaceOfferReply;
+import bisq.proto.grpc.PlaceOfferRequest;
+import bisq.proto.grpc.StopServerGrpc;
+import bisq.proto.grpc.StopServerReply;
+import bisq.proto.grpc.StopServerRequest;
 
 import io.grpc.Server;
 import io.grpc.ServerBuilder;
