+++ settings.gradle
@@ -1,4 +1,4 @@
-include 'protodefinition'
+include 'proto'
 include 'assets'
 include 'common'
 include 'p2p'
